# Improve uint128 performance by 10.8% and 7.9%