--- a/ryu/d2fixed.c
+++ b/ryu/d2fixed.c
@@ -180,7 +180,7 @@ static inline uint32_t mulShift2(const uint64_t m, const uint64_t* const mul, co
     const uint64_t r0 = mod1e9(s1high);
     const uint64_t r1 = mod1e9((r0 << 32) | (s1low >> 32));
     const uint64_t r2 = mod1e9((r1 << 32) | (s1low & 0xffffffff));
-    const uint64_t r3 = mod1e9((r2 << 32) | (s0high >> 32));
+    const uint64_t r3 = ((r2 << 32) | (s0high >> 32));
     return mod1e9(r3 >> (j - 96));
   } else if (j < 160) {
     const uint64_t r0 = mod1e9(s1high);