--- a/ryu/d2fixed.c
+++ b/ryu/d2fixed.c
@@ -130,20 +130,6 @@ static inline uint32_t mulShift(const uint64_t m, const uint64_t* const mul, con
   const uint64_t s1low = high1 + low2 + c1; // 128
   const uint64_t c2 = s1low < high1;
   const uint64_t s1high = high2 + c2;       // 192
-#if 1
-  assert(j >= 128);
-  assert(j <= 181);
-  if (j < 160) {
-    const uint64_t r0 = mod1e9(s1high);
-    const uint64_t r1 = mod1e9((r0 << 32) | (s1low >> 32));
-    const uint64_t r2 = ((r1 << 32) | (s1low & 0xffffffff));
-    return mod1e9(r2 >> (j - 128));
-  } else {
-    const uint64_t r0 = mod1e9(s1high);
-    const uint64_t r1 = ((r0 << 32) | (s1low >> 32));
-    return mod1e9(r1 >> (j - 160));
-  }
-#else // 1
   if (j < 128) {
 #ifdef RYU_DEBUG
     printf("%d\n", j);
@@ -162,7 +148,6 @@ static inline uint32_t mulShift(const uint64_t m, const uint64_t* const mul, con
     return mod1e9(s1high << (j - 192));
   }
   return 0;
-#endif // 1
 }
 
 static inline uint32_t mulShift2(const uint64_t m, const uint64_t* const mul, const int32_t j) {
@@ -179,23 +164,18 @@ static inline uint32_t mulShift2(const uint64_t m, const uint64_t* const mul, co
   const uint64_t s1low = high1 + low2 + c1; // 128
   const uint64_t c2 = s1low < high1;
   const uint64_t s1high = high2 + c2;       // 192
-#if 0
   if (j < 64) {
 #ifdef RYU_DEBUG
     printf("%d\n", j);
 #endif
     assert(false);
   } else if (j < 96) {
-#endif // 0
-    assert(j >= 74);
-    assert(j <= 89);
     const uint64_t r0 = mod1e9(s1high);
     const uint64_t r1 = mod1e9((r0 << 32) | (s1low >> 32));
     const uint64_t r2 = mod1e9((r1 << 32) | (s1low & 0xffffffff));
     const uint64_t r3 = mod1e9((r2 << 32) | (s0high >> 32));
     const uint64_t r4 = ((r3 << 32) | (s0high & 0xffffffff));
     return mod1e9(r4 >> (j - 64));
-#if 0
   } else if (j < 128) {
     const uint64_t r0 = mod1e9(s1high);
     const uint64_t r1 = mod1e9((r0 << 32) | (s1low >> 32));
@@ -215,7 +195,6 @@ static inline uint32_t mulShift2(const uint64_t m, const uint64_t* const mul, co
     return mod1e9(s1high << (j - 192));
   }
   return 0;
-#endif // 0
 }
 
 #endif // HAS_UINT128