--- a/ryu/d2fixed.c
+++ b/ryu/d2fixed.c
@@ -55,23 +55,54 @@ typedef __uint128_t uint128_t;
 #define POW10_ADDITIONAL_BITS 120
 
 #if defined(HAS_UINT128)
+static inline uint128_t umul256(const uint128_t a, const uint64_t bHi, const uint64_t bLo, uint128_t* const productHi) {
+  const uint64_t aLo = (uint64_t)a;
+  const uint64_t aHi = (uint64_t)(a >> 64);
+
+  const uint128_t b00 = (uint128_t)aLo * bLo;
+  const uint128_t b01 = (uint128_t)aLo * bHi;
+  const uint128_t b10 = (uint128_t)aHi * bLo;
+  const uint128_t b11 = (uint128_t)aHi * bHi;
+
+  const uint64_t b00Lo = (uint64_t)b00;
+  const uint64_t b00Hi = (uint64_t)(b00 >> 64);
+
+  const uint128_t mid1 = b10 + b00Hi;
+  const uint64_t mid1Lo = (uint64_t)(mid1);
+  const uint64_t mid1Hi = (uint64_t)(mid1 >> 64);
+
+  const uint128_t mid2 = b01 + mid1Lo;
+  const uint64_t mid2Lo = (uint64_t)(mid2);
+  const uint64_t mid2Hi = (uint64_t)(mid2 >> 64);
+
+  const uint128_t pHi = b11 + mid1Hi + mid2Hi;
+  const uint128_t pLo = ((uint128_t)mid2Lo << 64) + b00Lo;
+
+  *productHi = pHi;
+  return pLo;
+}
+
+// Returns the high 128 bits of the 256-bit product of a and b.
+static inline uint128_t umul256_hi(const uint128_t a, const uint64_t bHi, const uint64_t bLo) {
+  // Reuse the umul256 implementation.
+  // Optimizers will likely eliminate the instructions used to compute the
+  // low part of the product.
+  uint128_t hi;
+  umul256(a, bHi, bLo, &hi);
+  return hi;
+}
 
 // Unfortunately, gcc/clang do not automatically turn a 128-bit integer division
 // into a multiplication, so we have to do it manually.
-static inline uint128_t uint128_mod1e9(uint128_t v) {
-//  uint64_t m0 = (((uint64_t) (v >> 64)) % 1000000000) * 709551616;
-//  uint64_t s1 = m0 + (uint64_t) v;
-//  if (s1 < m0) {
-//    return (s1 % 1000000000 + 709551616) % 1000000000;
-//  } else {
-//    return s1 % 1000000000;
-//  }
+static inline uint32_t uint128_mod1e9(uint128_t v) {
+  // After multiplying, we're going to shift right by 29, then truncate to uint32_t.
+  // This means that we need only 29 + 32 = 61 bits, so we can truncate to uint64_t before shifting.
+  const uint64_t multiplied = (uint64_t) umul256_hi(v, 0x89705F4136B4A597u, 0x31680A88F8953031u);
 
-  uint64_t m0 = ((uint64_t) (v >> 64)) % 1000000000;
-  uint64_t m1 = ((uint64_t) v) % 1000000000;
-  // 2^64 % 1000000000 = 709551616
-  uint64_t m2 = (m0 * 709551616 + m1) % 1000000000;
-  return m2;
+  // For uint32_t truncation, see the mod1e9() comment in d2s_intrinsics.h.
+  const uint32_t shifted = (uint32_t) (multiplied >> 29);
+
+  return ((uint32_t) v) - 1000000000 * shifted;
 }
 
 // Best case: use 128-bit type.
@@ -90,7 +121,7 @@ static inline uint32_t mulShift(const uint64_t m, const uint64_t* const mul, con
   uint128_t s0 = b0 + (b1 << 64); // 0
   uint32_t c1 = s0 < b0;
   uint128_t s1 = b2 + (b1 >> 64) + c1; // 128
-  return (uint32_t) uint128_mod1e9(s1 >> (j - 128));
+  return uint128_mod1e9(s1 >> (j - 128));
 }
 
 #else // HAS_UINT128