These changes primarily improve mulShift(), activating the uint128 codepath for Clang/LLVM x64 targeting Windows, and improving uint128 perf (hopefully for Linux and Mac too) by 10.8% (for %f) and 7.9% (for %e). There's a small improvement from removing a special case for 0-shifting (previously removed for the non-uint128 codepaths), and a large improvement from implementing a 128-bit magic multiply.

Unlike the 64-bit magic multiplies, where we can just inspect Clang x64 codegen, I had to generate the 128-bit multiplier (`0x89705F4136B4A59731680A88F8953031`, written as two 64-bit constants) and shift value (`29`) myself, with the classic algorithm in Hacker's Delight. (The "add" indicator was 0, resulting in simpler code, happily.) The new helper functions `umul256()` and `umul256_hi()` are modified copies of Ryu's `umul128()` and `umulh()` in d2s_intrinsics.h.

For correctness testing, I've tested over 500 billion unique random doubles with MSVC x64 (non-uint128) and over 568 billion with LLVM uint128, for both %f and %e (compared to MSVC UCRT `sprintf_s()`), with zero differences.

Compiler    | Before  | After
------------|---------|--------
MSVC x86 %f | 361.681 | 361.322
LLVM x86 %f | 352.800 | 352.596
MSVC x64 %f | 159.763 | 159.702
LLVM x64 %f | 116.678 | 116.277
LLVM 128 %f | 112.508 | 101.502

Compiler    | Before  | After
------------|---------|--------
MSVC x86 %e | 114.038 | 113.568
LLVM x86 %e | 104.484 | 104.387
MSVC x64 %e |  56.020 |  55.130
LLVM x64 %e |  40.340 |  40.147
LLVM 128 %e |  40.272 |  37.333

Individual commits:

---

Allow Clang/LLVM to use uint128 on Windows.

It's somewhat faster than the intrinsics codepath.

Support for uint128 on Windows is definitely incomplete (attempting to use division or modulo triggers linker errors due to unimplemented library helpers), but I haven't encountered any problems with the current code.

Add RYU_AVOID_UINT128 as an escape hatch (for perf testing).

---

Revert "Restrict mulShift() and mulShift2() to their actual ranges."

This reverts commit 879d2b76f4f14531d4c9bf6ff610e513976401db.

---

Fix another typo-bug in unused code.

mulShift() and mulShift2() perform a wide multiplication followed by a right shift and a modulo.

The last case, for j within [192, 255], which shifts away everything except for the s1high part of the wide mult, was incorrectly performing a left shift.

This was especially easy to miss because the other cases need `<< 32`.

This didn't affect correctness because the branch isn't being taken (currently `j <= 180` holds), but this should be fixed just in case.

---

Fix yet another typo-bug in unused code.

mulShift2() was incorrectly performing an extra mod1e9() in one case, unlike every other case in mulShift() and mulShift2().

This didn't affect correctness; previously, the broken case wasn't activated due to the possible range of j, and now mulShift2() is completely unused.

(I'm fixing this in case the code is ever needed again.)

---

Remove mulShift2() which is now unused.

---

d2fixed_full_table.h needs to include stdint.h.

---

Improve d2fixed_buffered_n() perf for `e2 == -53`.

We're printing `m2 * 2^e2`, and this branch is for the integer part (to the left of the decimal point).

m2 is the ieeeMantissa with the implicit bit restored (0 for subnormal, 1 for normal), so it's a 53-bit number (at most).

When `e2 == -53`, `m2 * 2^-53` has no integer part, so this branch doesn't need to be taken. (The following code will write a '0' for the integer part, as it already does for more negative exponents.)

This is a strict performance improvement for this exponent, and now this test is consistent with d2exp_buffered_n().

As this affects only one exponent, it's unobservable in the benchmark.

---

Improve mulShift() comments.

After commit c898a78, mulShift() is always called with a shift in the range [128, 180]. I'm updating the comments accordingly.

---

Add an assertion to mulShift().

The code is taking a shortcut when determining whether to carry out of the `high1 + low2 + c1` additions. As far as I can tell, this shortcut is valid given the mantissa range and static tables being used, but there should be an assertion just in case.

---

Store carry bits in uint32_t.

They're either 0 or 1, so at best this can help compilers realize that they don't need to do full 128 + 128 or 64 + 64 additions. At worst, this is a stylistic change to indicate the narrow range.

(This turned out to be unobservable in the benchmark.)

---

Restrict mulShift() to its updated range.

This is a strict performance improvement and reduces codegen size.

---

Don't avoid 0-shifts in uint128 mulShift().

The benchmark indicates that attempting to avoid 0-shifting is counterproductive.

---

Improve uint128 perf with a magic multiply.

---

Use DIGIT_TABLE more often.

This appears to be a very slight performance improvement.